requirements.txt
argparse==1.2.1
backports.ssl-match-hostname==3.4.0.2
markdown2==2.2.1
py-bcrypt==0.4
sh==1.09
tornado==3.2
wsgiref==0.1.2
